+++ cmd/gowdk/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/cssbruno/gowdk/addons/ssr"
 )
 
-const version = "0.2.7"
+const version = "0.2.8"
 
 var (
 	defaultSourceIncludes = []string{"**/*.gwdk"}